requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# LangChain & LangGraph (versiones modernas)
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.6.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Vector store
chromadb>=0.4.22

# PDF generation
weasyprint>=60.2
jinja2>=3.1.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.10.0
requests>=2.32.0

